services/stack/services/rabbit-worker/requirements.txt
pika
mysql-connector-python
